+++ frontend/src/components/Card.js
@@ -3,7 +3,6 @@ import React, { useState } from "react";
 import axios from "axios";
 import CONFIG from "../config";
 import { logFrontendEvent } from "../utils/logger";
-import moment from "moment-timezone";
 import { formatTimestamp } from "../utils/time";  
 
 function Card({ task, refreshTasks, user }) {